+++ py3/dispy.py
@@ -634,6 +634,12 @@ class _Cluster(object, metaclass=MetaSingleton):
             self.worker_thread = threading.Thread(target=self.worker)
             self.worker_thread.daemon = True
             self.worker_thread.start()
+            # Under Windows dispynode may send objects with
+            # '__mp_main__' scope, so make an alias to '__main__'.
+            # TODO: Make alias even if client is not Windows? It is
+            # possible the client is not Windows, but a node is.
+            if os.name == 'nt' and '__mp_main__' not in sys.modules:
+                sys.modules['__mp_main__'] = sys.modules['__main__']
 
     def udp_server(self, coro=None):
         # generator
@@ -1123,10 +1129,9 @@ class _Cluster(object, metaclass=MetaSingleton):
                 self.poll_interval = num_min(self.poll_interval, cluster.poll_interval)
             if self.poll_interval:
                 self.timer_coro.resume(True)
-            raise StopIteration
         # if a node is added with 'allocate_node', compute is already
         # initialized, so don't reinitialize it
-        if compute.id is None:
+        elif compute.id is None:
             compute.id = self.compute_id
             self.compute_id += 1
             self._clusters[compute.id] = cluster
@@ -1144,21 +1149,21 @@ class _Cluster(object, metaclass=MetaSingleton):
                 self.poll_interval = num_min(self.poll_interval, cluster.poll_interval)
             if self.pulse_interval or self.ping_interval or self.poll_interval:
                 self.timer_coro.resume(True)
-
-        yield self.send_ping_cluster(cluster, coro=coro)
-        compute_nodes = []
-        for ip_addr, node in self._nodes.items():
-            if compute.id in node.clusters:
-                continue
-            for node_alloc in cluster._node_allocs:
-                cpus = node_alloc.allocate(cluster, node.ip_addr, node.name, node.avail_cpus)
-                if cpus <= 0:
+        else:
+            yield self.send_ping_cluster(cluster, coro=coro)
+            compute_nodes = []
+            for ip_addr, node in self._nodes.items():
+                if compute.id in node.clusters:
                     continue
-                node.cpus = min(node.avail_cpus, cpus)
-                cluster._dispy_nodes.pop(node.ip_addr, None)
-                compute_nodes.append(node)
-        for node in compute_nodes:
-            yield self.setup_node(node, [compute], coro=coro)
+                for node_alloc in cluster._node_allocs:
+                    cpus = node_alloc.allocate(cluster, node.ip_addr, node.name, node.avail_cpus)
+                    if cpus <= 0:
+                        continue
+                    node.cpus = min(node.avail_cpus, cpus)
+                    cluster._dispy_nodes.pop(node.ip_addr, None)
+                    compute_nodes.append(node)
+            for node in compute_nodes:
+                yield self.setup_node(node, [compute], coro=coro)
 
     def del_cluster(self, cluster, coro=None):
         # generator
@@ -1675,12 +1680,12 @@ class _Cluster(object, metaclass=MetaSingleton):
             # TODO: need to check all clusters are deleted?
             self.shelf.close()
             self.shelf = None
-            for ext in ('', '.db'):
-                try:
-                    os.remove(self.recover_file + ext)
-                    break
-                except:
-                    pass
+            for ext in ('', '.db', '.bak', '.dat', '.dir'):
+                if os.path.isfile(self.recover_file + ext):
+                    try:
+                        os.remove(self.recover_file + ext)
+                    except:
+                        pass
 
 
 class JobCluster(object):
@@ -2561,12 +2566,12 @@ def recover_jobs(recover_file, timeout=None, terminate_pending=False):
             logger.warning('invalid key "%s" ignored' % key)
     shelf.close()
     if not cluster or not computes or not shelf_nodes:
-        for ext in ('', '.db'):
-            try:
-                os.remove(recover_file + ext)
-                break
-            except:
-                pass
+        for ext in ('', '.db', '.bak', '.dat', '.dir'):
+            if os.path.isfile(recover_file + ext):
+                try:
+                    os.remove(recover_file + ext)
+                except:
+                    pass
         return []
 
     nodes = {}
@@ -2695,12 +2700,12 @@ def recover_jobs(recover_file, timeout=None, terminate_pending=False):
     asyncoro_scheduler.finish()
 
     if pending['count'] == 0 and pending['resend_req_done'] is True:
-        for ext in ('', '.db'):
-            try:
-                os.remove(recover_file + ext)
-                break
-            except:
-                pass
+        for ext in ('', '.db', '.bak', '.dat', '.dir'):
+            if os.path.isfile(recover_file + ext):
+                try:
+                    os.remove(recover_file + ext)
+                except:
+                    pass
     return pending['jobs']
